backend/compliance/requirements.txt
flask
flask_sqlalchemy
mysql-connector-python
sqlalchemy
requests
python-dotenv

